pyproject.toml
[project]
name = "srenity"
version = "0.1.0"
description = "SRE Workflow Agent using LangGraph for incident management and log analysis"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langgraph",
    "langchain",
    "langchain-openai",
    "langchain-anthropic",
    "pandas",
    "fastapi",
    "uvicorn",
    "pydantic",
    "httpx",
    "python-multipart",
    "python-jose[cryptography]",
    "passlib[bcrypt]",
    "aiofiles",
    "numpy",
    "matplotlib",
    "seaborn",
    "scikit-learn",
    "loguru",
    "typing-extensions",
    "python-dotenv",
    "opik",
    "psutil",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "black",
    "ruff",
    "mypy",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.black]
line-length = 88
target-version = ['py312']

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true 
